Add test for async sample retrieval
The test tries to mimic a common pattern, which is the processing of incoming
subscription data via an async function. The resulting Future shall be
compatible with the spawn function of common executors (tokio in this
case). This means that the Future itself and its output are Send +
'static.

The test requires fixing the lifetime bounds of receive. In addition, it
is (for now) not possible to use a trait for the receiving container.
Instead, a concrete class is used.